+++ Modules/SWMB/LocalMachine-Network.psm1
@@ -408,6 +408,39 @@ Function TweakViewAnonymousShareAccess { # RESINFO
 
 ################################################################
 
+# Anonymous SID/Name translation must not be allowed
+# W11 STIG V-253452 https://www.stigviewer.com/stigs/microsoft-windows-11-security-technical-implementation-guide/2025-05-15/finding/V-253452
+
+# Disable
+Function TweakDisableAnonymousNameTranslation { # RESINFO
+	Write-Output "Disabling Anonymous SID/Name translation ..."
+	$RegPath = 'HKLM:\SYSTEM\CurrentControlSet\Control\Lsa'
+	Remove-ItemProperty -Path $RegPath -Name "TurnOffAnonymousBlock" -ErrorAction SilentlyContinue
+}
+
+# Enable
+Function TweakEnableAnonymousNameTranslation { # RESINFO
+	Write-Output "Enabling Anonymous SID/Name translation ..."
+	$RegPath = 'HKLM:\SYSTEM\CurrentControlSet\Control\Lsa'
+	Set-ItemProperty -Path $RegPath -Name "TurnOffAnonymousBlock" -Type DWord -Value 1
+}
+
+# View
+Function TweakViewAnonymousNameTranslation { # RESINFO
+	Write-Output "Viewing Anonymous SID/Name translation  (0 or not exist: Disable (Default, Recommanded), 1: Enable)..."
+	$RegPath = 'HKLM:\SYSTEM\CurrentControlSet\Control\Lsa'
+	$RegFields = @{
+		TurnOffAnonymousBlock = @{
+			OkValues = @(0, $Null)   # 0 or not exist
+			Description = "Disable anonymous SID/Name translation "
+			Remediation = "DisableAnonymousNameTranslation (W11 STIG V-253452)"
+		}
+	}
+	SWMB_GetRegistrySettings -Path $RegPath -Rules $RegFields | SWMB_WriteSettings
+}
+
+################################################################
+
 # Remote Assistance and Solicited Remote Assistance
 # Remote assistance allows another user to view or take control of the local session of a user.
 # Disable Remote Assistance - Not applicable to Server (unless Remote Assistance is explicitly installed)